+++ src/functions.php
@@ -10,6 +10,7 @@ use GreatMarketrealmExpansions\Content\Types\ContentTypeCatalogue;
 use GreatMarketrealmExpansions\Expansions\ExpansionRegistry;
 use GreatMarketrealmExpansions\Expansions\Loading\ExpansionFileLoader;
 use GreatMarketrealmExpansions\Integration\Bridge;
+use GreatMarketrealmExpansions\Import\ImportService;
 use GreatMarketrealmExpansions\Library\Library;
 use GreatMarketrealmExpansions\Rules\RuleEngine;
 
@@ -74,4 +75,11 @@ function library(): Library
     $kernel = Kernel::instance();
     $kernel->boot();
     return $kernel->library();
+}
+
+function importer(): ImportService
+{
+    $kernel = Kernel::instance();
+    $kernel->boot();
+    return $kernel->importer();
 }